feat : BaseEntity추가
