+++ urecating/src/main/java/com/uting/urecating/domain/BaseEntity.java
@@ -1,0 +1,33 @@
+package com.uting.urecating.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Setter
+@Getter
+@MappedSuperclass
+public class BaseEntity {
+
+    @Column(name = "created_at", updatable = false)
+    private LocalDateTime createdAt; //생성날짜
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt; //수정날짜
+
+    @PrePersist
+    public void prePersist(){
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = this.createdAt;
+    }
+
+    @PreUpdate
+    public void onPreUpdate(){
+        this.updatedAt = LocalDateTime.now();
+    }
+}
